scx_bpfland: prevent stall with sched_setaffinity()
Dispatching a task to a per-CPU DSQ while it's blocked in
__set_cpus_allowed_ptr() seems to lock the DSQ completely, meaning that
the dispatched tasks can't be consumed by the associated CPU (but they
can be consumed by other CPUs).

As a workaround, automatically bounce tasks to the shared DSQ when they
are dispatched from sched_setaffinity() to prevent the stall condition.

This is similar to commit 86db45f ("scx_rustland_core: prevent deadlock
with per-CPU DSQs and CPU affinity") in scx_rustland_core.

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>